+++ relayer/relayer.go
@@ -6,6 +6,8 @@ import (
 	"math/big"
 	"time"
 
+	"github.com/celestiaorg/orchestrator-relayer/helpers"
+
 	coregethtypes "github.com/ethereum/go-ethereum/core/types"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
@@ -26,6 +28,7 @@ type Relayer struct {
 	P2PQuerier *p2p.Querier
 	EVMClient  *evm.Client
 	logger     tmlog.Logger
+	Retrier    *helpers.Retrier
 }
 
 func NewRelayer(
@@ -34,6 +37,7 @@ func NewRelayer(
 	p2pQuerier *p2p.Querier,
 	evmClient *evm.Client,
 	logger tmlog.Logger,
+	retrier *helpers.Retrier,
 ) (*Relayer, error) {
 	return &Relayer{
 		TmQuerier:  tmQuerier,
@@ -41,6 +45,7 @@ func NewRelayer(
 		P2PQuerier: p2pQuerier,
 		EVMClient:  evmClient,
 		logger:     logger,
+		Retrier:    retrier,
 	}, nil
 }
